requirements.txt
flask==3.1.0
flask-cors==5.0.0
ultralytics>=8.0.0
Pillow>=10.4.0
pillow-heif>=0.13.0
numpy>=1.26.0
opencv-python>=4.8.0
shapely>=2.0.6
